refactor: extract shared.py from mail _legacy module
